# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: CRM module for Uralintern
  description: API documentation for Uralintern CRM Module
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: kayakto server url
security:
  - BearerAuth: []
...
paths:
...
  /messages/update-message:
    put:
      tags:
        - message-controller
      operationId: updateMessage
      description: "Обновление шаблона сообщения"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageUpdateRequest'
        required: true
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /events-students/{eventId}/send/{studentId}:
    put:
      tags:
        - event-student-controller
      operationId: sendStudentToEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-students/{eventId}/reject/{studentId}:
    put:
      tags:
        - event-student-controller
      operationId: rejectStudentRequest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-students/{eventId}/accept/{studentId}:
    put:
      tags:
        - event-student-controller
      operationId: acceptStudentRequest
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-students/change-curator/{eventId}/students/{studentId}/curator/{newCuratorId}:
    put:
      tags:
        - event-student-controller
      operationId: changeStudentCurator
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: newCuratorId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /notifications/user/{userId}:
    get:
      tags:
        - notification-controller
      operationId: getUserNotifications
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /messages/{messageId}:
    get:
      tags:
        - message-controller
      operationId: getMessageById
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /messages/event/{eventId}:
    get:
      tags:
        - message-controller
      operationId: getEventMessages
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /events-students/{eventId}/waiting-students:
    get:
      tags:
        - event-student-controller
      operationId: getSentStudentsInfo
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-students/{eventId}/students:
    get:
      tags:
        - event-student-controller
      operationId: getStudentsInfo
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-students/{eventId}/student-status/{studentId}:
    get:
      tags:
        - event-student-controller
      operationId: getStudentStatus
      parameters:
        - name: eventId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-students/{eventId}/student-can-send/{studentId}:
    get:
      tags:
        - event-student-controller
      operationId: canSend
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /events-students/{eventId}/accepted-students:
    get:
      tags:
        - event-student-controller
      operationId: getAcceptedStudentsInfo
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /events-students/{eventId}/delete/{studentId}:
    delete:
      tags:
        - event-student-controller
      operationId: deleteStudentFromEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  responses:
      OkResponse:
        description: OK, запрос выполнен успешно
        content:
          '*/*':
            schema:
              type: boolean
      BadRequestResponse:
        description: Bad Request, запрос неверный со стороны клиента
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
      NotFoundResponse:
        description: Not Found, такого элемента не существует
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
      InternalServerErrorResponse:
        description: Internal Server Error, запрос вызвал ошибку на сервере
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "User with email email123@gmail.com not found"
...
    MessageUpdateRequest:
      type: object
      properties:
        messageId:
          type: integer
          format: int64
          example: 40
        text:
          type: string
          example: "Вы приняты на мероприятие \"Продажа часов\". Присоединяйтесь к чату: t.me/chatBitrix"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT